feat(#690): move zone_helpers to lib/, delete user_helpers re-export shim
zone_helpers.py has zero kernel dependencies — it only uses
rebac_manager (passed as Any). Belongs in tier-neutral lib/.

Also:
- Delete server/auth/user_helpers.py (was just a re-export shim)
- Update all callers to import directly from canonical locations:
  - Zone functions: nexus.lib.zone_helpers
  - User queries: nexus.auth.user_queries
- Update test assertions and logger name references
- 12 files changed
