+++ tests/unit/core/test_zone_boundary_security.py
@@ -16,8 +16,8 @@ from nexus import LocalBackend, NexusFS
 from nexus.contracts.types import OperationContext
 from nexus.core.config import ParseConfig, PermissionConfig
 from nexus.factory import create_nexus_fs
+from nexus.lib.zone_helpers import add_user_to_zone
 from nexus.rebac.permissions_enhanced import AdminCapability
-from nexus.server.auth.user_helpers import add_user_to_zone
 from nexus.storage.raft_metadata_store import RaftMetadataStore
 from nexus.storage.record_store import SQLAlchemyRecordStore
